[AppsLogsEventSource] Move and minor refactor
